Bring code up-to-date, mostly.
Diagrams is now at version 1.4, from version 1.0.

The command line arguments for euler angles for the
legend is disabled because the parsing is more complicated.
The unit tests for the parser are also commented out.

The angle fudging in the legend had to change and now is
symmetrical, which may be a clue as to why it's needed.

The right endpoint of the orange arc in the ellipse
output is closer to where it should be but still a tiny
bit too far to the right.